feat: add basic context to plop and refactor plop generator
