+++ plop/air-context/index.ts.hbs
@@ -1,0 +1,4 @@
+export * from './{{pascalCase name}}.context';
+export * from './{{pascalCase name}}.types';
+export * from './{{pascalCase name}}.hooks';
+export * from './{{pascalCase name}}.reducer';
