+++ plop/context.generator.js
@@ -1,0 +1,44 @@
+module.exports = function (plop) {
+  plop.setGenerator('context', {
+    description: 'Basic React Context',
+    prompts: [
+      {
+        type: 'input',
+        name: 'name',
+        message: 'Context name:',
+      },
+      {
+        type: 'input',
+        name: 'module',
+        message: 'Module name (leave empty if context is common)',
+      },
+    ],
+    actions: (data) => {
+      let path = '';
+
+      if (data && data.module) {
+        path = `../src/modules/${data.module}/contexts`;
+      } else {
+        path = `../src/contexts`;
+      }
+
+      return [
+        {
+          type: 'add',
+          path: `${path}/{{pascalCase name}}/index.ts`,
+          templateFile: './context/index.ts.hbs',
+        },
+        {
+          type: 'add',
+          path: `${path}/{{pascalCase name}}/{{pascalCase name}}.context.tsx`,
+          templateFile: './context/context.tsx.hbs',
+        },
+        {
+          type: 'add',
+          path: `${path}/{{pascalCase name}}/{{pascalCase name}}.hooks.tsx`,
+          templateFile: './context/hooks.ts.hbs',
+        },
+      ];
+    },
+  });
+};
