+++ plop/context/context.tsx.hbs
@@ -1,86 +1,43 @@
-import { createContext, useMemo, useReducer } from 'react';
-import { {{pascalCase name}}ActionsType, {{pascalCase name}}StateType } from './{{pascalCase name}}.types';
-import { {{camelCase name}}Reducer } from './{{pascalCase name}}.reducer';
+import { createContext, useMemo, useState } from 'react';
 
-/**
- *
- * Separated into two contexts to avoid unnecessary re-renders
- * when only one of the contexts changes.
- *
- * See: https://kentcdodds.com/blog/how-to-optimize-your-context-value
- */
-export const {{pascalCase name}}State = createContext<{{pascalCase name}}StateType>({} as {{pascalCase name}}StateType);
-export const {{pascalCase name}}Actions = createContext<{{pascalCase name}}ActionsType>({} as {{pascalCase name}}ActionsType);
+export type {{pascalCase name}}Type = {
+  count: number;
+  setCount: (count: number) => void;
+};
 
 /**
  * Initial state
  *
  * This is used to initialize the state of the context.
  */
-const INITIAL_STATE = {
+const INITIAL_STATE: {{pascalCase name}}Type = {
   count: 0,
+  setCount: () => {},
 };
 
 /**
+ * Context
  *
- * {{pascalCase name}}Provider
- *
- * This is the provider for the {{pascalCase name}}.
+ * This is the context that is used to store the state and the dispatcher.
  */
+export const {{pascalCase name}} = createContext<{{pascalCase name}}Type>(INITIAL_STATE);
+
 export function {{pascalCase name}}Provider({
   children,
   initialState = INITIAL_STATE,
 }: {
   children: React.ReactNode;
-  initialState: {{pascalCase name}}StateType;
+  initialState: {{pascalCase name}}Type;
 }) {
-  /**
-   * useReducer
-   *
-   * Context state reducer
-   */
-  const [state, dispatch] = useReducer({{camelCase name}}Reducer, {
-    ...initialState,
-  });
+  const [count, setCount] = useState(initialState.count);
 
-  /**
-   *
-   * customAction
-   *
-   * This is an example of a custom action.
-   */
-  const customAction = async (payload: number) => {
-    dispatch({ type: 'set', payload });
-    await new Promise((resolve) => {
-      setTimeout(resolve, 1000);
-    });
-    dispatch({ type: 'increment' });
-  };
-
-  /**
-   *
-   * actionsValue
-   *
-   * This is the value of the actions context.
-   */
-  const actionsValue = useMemo(
-    (): {{pascalCase name}}ActionsType => ({
-      increment: () => dispatch({ type: 'increment' }),
-      set: (payload: number) => dispatch({ type: 'set', payload }),
-      customAction,
+  const value = useMemo(
+    (): {{pascalCase name}}Type => ({
+      count,
+      setCount,
     }),
-    [],
+    [count, setCount],
   );
 
-  /**
-   *
-   * This is the value of the state context.
-   *
-   * See: https://kentcdodds.com/blog/how-to-optimize-your-context-value
-   */
-  return (
-    <{{pascalCase name}}State.Provider value={state}>
-      <{{pascalCase name}}Actions.Provider value={actionsValue}>{children}</{{pascalCase name}}Actions.Provider>
-    </{{pascalCase name}}State.Provider>
-  );
+  return <{{pascalCase name}}.Provider value={value}>{children}</{{pascalCase name}}.Provider>;
 }
