+++ plop/context/index.ts.hbs
@@ -1,4 +1,2 @@
 export * from './{{pascalCase name}}.context';
-export * from './{{pascalCase name}}.types';
 export * from './{{pascalCase name}}.hooks';
-export * from './{{pascalCase name}}.reducer';
